+++ src/main/java/wootecamp/chess/game/ConsoleGameInputManager.java
@@ -1,4 +1,4 @@
-package wootecamp.game;
+package wootecamp.chess.game;
 
 import java.util.Scanner;
