+++ src/main/java/wootecamp/chess/game/Game.java
@@ -1,23 +1,23 @@
-package wootecamp.game;
+package wootecamp.chess.game;
 
 import wootecamp.chess.board.Board;
 import wootecamp.chess.board.BoardPosition;
 import wootecamp.chess.board.MoveVector;
+import wootecamp.chess.game.state.EndState;
+import wootecamp.chess.game.state.ReadyState;
+import wootecamp.chess.game.state.State;
 import wootecamp.chess.pieces.Direction;
 import wootecamp.chess.pieces.Piece;
-import wootecamp.game.state.EndState;
-import wootecamp.game.state.ReadyState;
-import wootecamp.game.state.State;
 
 public class Game {
     private final GameInputManager gameInputManager;
     private final GameOutputManager gameOutputManager;
     private final Board board;
-    
+
 
     private State state = new ReadyState(this);
     private Piece.Color curTurnColor = Piece.Color.WHITE;
-    
+
     public Game(GameInputManager gameInputManager, GameOutputManager gameOutputManager, Board board) {
         this.gameInputManager = gameInputManager;
         this.gameOutputManager = gameOutputManager;
@@ -35,7 +35,7 @@ public class Game {
 
     public void move(BoardPosition source, BoardPosition target) {
         Piece piece = board.findPiece(source);
-        if(piece.getColor() != curTurnColor) {
+        if (piece.getColor() != curTurnColor) {
             gameOutputManager.showError("차례가 아닙니다.");
             gameOutputManager.showBoard(board);
             return;
@@ -107,7 +107,6 @@ public class Game {
     }
 
     public boolean isEnd() {
-        //TODO 구조 개선
         return state instanceof EndState;
     }
