+++ src/main/java/wootecamp/chess/game/state/EndState.java
@@ -1,6 +1,6 @@
-package wootecamp.game.state;
+package wootecamp.chess.game.state;
 
-import wootecamp.game.Game;
+import wootecamp.chess.game.Game;
 
 public class EndState extends State {
     public EndState(Game game) {
