+++ src/test/java/wootecamp/chess/board/BoardTest.java
@@ -1,11 +1,10 @@
-package wootecamp.chess;
+package wootecamp.chess.board;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import wootecamp.chess.board.Board;
-import wootecamp.chess.board.BoardPosition;
 import wootecamp.chess.pieces.Piece;
+import wootecamp.chess.pieces.PieceComparator;
 import wootecamp.chess.pieces.PieceFactory;
 
 import java.util.List;
