+++ src/test/java/wootecamp/chess/board/RankTest.java
@@ -1,8 +1,7 @@
-package wootecamp.chess;
+package wootecamp.chess.board;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import wootecamp.chess.board.Rank;
 import wootecamp.chess.pieces.Piece;
 
 import static org.assertj.core.api.Assertions.assertThat;
